+++ src/modules/ce/components/opportunity/OpportunityBanner.tsx
@@ -4,8 +4,8 @@ import React, { useMemo } from 'react';
 import ButtonLink from '@/components/elements/ButtonLink';
 import { CommonLabeledTextBlock } from '@/components/elements/CommonLabeledTextBlock';
 import { useIsMobile } from '@/hooks/useIsMobile';
-import BeginReferralButton from '@/modules/ce/components/BeginReferralButton';
-import ReferralStatusChip from '@/modules/ce/components/ReferralStatusChip';
+import BeginReferralButton from '@/modules/ce/components/opportunity/BeginReferralButton';
+import ReferralStatusChip from '@/modules/ce/components/referral/ReferralStatusChip';
 import { clientNameFromRecordWithOptionalClient } from '@/modules/hmis/hmisUtil';
 import { useProjectDashboardContext } from '@/modules/projects/components/ProjectDashboard';
 import { ProjectDashboardRoutes } from '@/routes/routes';
