+++ src/modules/ce/components/project/ProjectReferralsTable.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import CommonTruncatedList from '@/components/elements/CommonTruncatedList';
 import { ColumnDef } from '@/components/elements/table/types';
 import useSafeParams from '@/hooks/useSafeParams';
-import ReferralStatusChip from '@/modules/ce/components/ReferralStatusChip';
+import ReferralStatusChip from '@/modules/ce/components/referral/ReferralStatusChip';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 import {
   clientNameFromRecordWithOptionalClient,
