+++ src/modules/ce/components/referral/ReferralSteps.tsx
@@ -2,9 +2,9 @@ import { Divider, Paper, Stack, Typography } from '@mui/material';
 import React, { useMemo } from 'react';
 import NotFound from '@/components/pages/NotFound';
 import useSearchParamsState from '@/hooks/useSearchParamState';
-import { useReferralContext } from '@/modules/ce/components/ReferralPage';
-import ReferralStepCard from '@/modules/ce/components/ReferralStepCard';
-import ReferralWayfinder from '@/modules/ce/components/ReferralWayfinder';
+import { useReferralContext } from '@/modules/ce/components/referral/ReferralPage';
+import ReferralStepCard from '@/modules/ce/components/referral/ReferralStepCard';
+import ReferralWayfinder from '@/modules/ce/components/referral/ReferralWayfinder';
 import { CeReferralStepStatus } from '@/types/gqlTypes';
 
 interface Props {}
