+++ src/routes/protected.tsx
@@ -53,11 +53,11 @@ import EnrollmentCaseNotes from '@/modules/caseNotes/components/EnrollmentCaseNo
 
 import AdminCoordinatedEntry from '@/modules/ce/components/admin/AdminCoordinatedEntry';
 import ClientReferralsPage from '@/modules/ce/components/client/ClientReferralsPage';
-import Opportunity from '@/modules/ce/components/Opportunity';
-import ProjectCePage from '@/modules/ce/components/ProjectCePage';
-import ReferralPage from '@/modules/ce/components/ReferralPage';
-import ReferralStep from '@/modules/ce/components/ReferralStep';
-import ReferralSteps from '@/modules/ce/components/ReferralSteps';
+import Opportunity from '@/modules/ce/components/opportunity/Opportunity';
+import ProjectCePage from '@/modules/ce/components/project/ProjectCePage';
+import ReferralPage from '@/modules/ce/components/referral/ReferralPage';
+import ReferralStep from '@/modules/ce/components/referral/ReferralStep';
+import ReferralSteps from '@/modules/ce/components/referral/ReferralSteps';
 import ClientDashboard from '@/modules/client/components/pages/ClientDashboard';
 import ClientProfilePage from '@/modules/client/components/pages/ClientProfilePage';
 import CreateClientPage from '@/modules/client/components/pages/CreateClientPage';
